+++ src/shared/components/index.ts
@@ -1,5 +1,6 @@
 export * from './AppInfo';
 export * from './AppSidebar';
 export * from './ButtonCounter';
+export * from './Highlight';
 export * from './Markdown';
 export * from './AppBreadcrumb';