+++ src/showcase/examples/AppBarPage.tsx
@@ -1,42 +1,25 @@
-import React, { Fragment, useEffect, useState } from 'react';
-import { AppBar, AppBarTitle, COLOR, FormControl, HomeSolidIcon, IAppBarProps, IconButton, Tab, Tabset } from '../../lib';
+import React, { Fragment, useEffect } from 'react';
+import { AppBar, AppBarTitle, COLOR, FormControl, HomeSolidIcon, IAppBarProps, IconButton } from '../../lib';
 import { IShowcaseBaseProps, withOptions } from './components';
-import { fileLoaderService, Markdown } from '../../shared';
 
 const AppBarPageBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<IAppBarProps>) => {
 
-	const [markdown, setMarkdown] = useState('');
-
 	useEffect(() => {
-		init();
-
 		setSettingsControls({
-			color: new FormControl(settingValues.color, [], 'select', { label: 'color', options: [{ value: COLOR.primary }, { value: COLOR.light }] }),
+			color: new FormControl(settingValues.color, [], 'select', { label: 'color', options: [{ label: 'primary', value: COLOR.primary }, { label: 'light', value: COLOR.light }] }),
 			shadow: new FormControl(settingValues.shadow, [], 'checkbox', { label: 'Shadow' }),
 		});
 	}, []);
 
-	const init = async () => {
-		const response = await fileLoaderService.get<string>('./showcase/appbar.md');
-		setMarkdown(response.data);
-	}
-
 	return (
 		<Fragment>
-			<Tabset>
-				<Tab eventKey="tab1" title="Preview">
-					<AppBar
-						color={settingValues.color}
-						shadow={settingValues.shadow}
-					>
-						<AppBarTitle>Navbar</AppBarTitle>
-						<IconButton className="ml-auto" color={COLOR.light} icon={<HomeSolidIcon />} />
-					</AppBar>
-				</Tab>
-				<Tab eventKey="tab2" title="Usage">
-					<Markdown text={markdown} />
-				</Tab>
-			</Tabset>
+			<AppBar
+				color={settingValues.color}
+				shadow={settingValues.shadow}
+			>
+				<AppBarTitle>Navbar</AppBarTitle>
+				<IconButton className="ml-auto" color={COLOR.light} icon={<HomeSolidIcon />} />
+			</AppBar>
 		</Fragment>
 	);
 }